# reduce GQA test combinations